small fixes to GR-Verified Reports column on profiles

+++ src/api/export/users/queries/profiles.js
@@ -179,7 +179,7 @@ function handleGetProfiles(req, res) {
                     verifiedFollowerCount: record.get('verifiedFollowerCount') ? parseInt(record.get('verifiedFollowerCount').toString()) : 0,
                     nip56_totalGrapeRankScore: record.get('nip56_totalGrapeRankScore') ? parseFloat(record.get('nip56_totalGrapeRankScore').toString()) : null,
                     nip56_totalReportCount: record.get('nip56_totalReportCount') ? parseInt(record.get('nip56_totalReportCount').toString()) : null,
-                    nip56_totalVerifiedReportCount: record.get('nip56_totalVerifiedReportCount') ? parseInt(record.get('nip56_totalVerifiedReportCount').toString()) : null
+                    nip56_totalVerifiedReportCount: record.get('nip56_totalVerifiedReportCount') ? parseFloat(record.get('nip56_totalVerifiedReportCount').toString()) : null
                   };
                 });
